Crate substrate_cli

Substrate CLI library.

Modules

error
Initialization errors.

informant
Console informant. Prints sync progress and block events. Runs on the calling thread.

Macros

impl_augment_clap
Macro for implementing the AugmentClap trait. This requires that the given type uses derive(StructOpt)!

Structs

NoCustom
A special commandline parameter that expands to nothing. Should be used as custom subcommand/run arguments if no custom values are required.

SharedParams
Shared parameters used by all CoreParams.

VersionInfo
Executable version. Used to pass version information from the root crate.

Enums

CoreParams
All core commands that are provided by default.

Traits

AugmentClap
Something that can augment a clap app with further parameters. derive(StructOpt) is implementing this function by default, so a macro impl_augment_clap! is provided to simplify the implementation of this trait.

GetLogFilter
Returns the log filter given by the user as commandline argument.

IntoExit
Something that can be converted into an exit signal.

Functions

create_config_with_db_path
Creates a configuration including the database path.

parse_and_execute
Parse command line interface arguments and executes the desired command.