Function substrate_cli::parse_and_execute

pub fn parse_and_execute<'a, F, CC, RP, S, RS, E, I, T>(
    spec_factory: S, 
    version: &VersionInfo, 
    impl_name: &'static str, 
    args: I, 
    exit: E, 
    run_service: RS
) -> Result<Option<CC>> where
    F: ServiceFactory,
    S: FnOnce(&str) -> Result<Option<ChainSpec<FactoryGenesis<F>>>, String>,
    CC: StructOpt + Clone + GetLogFilter,
    RP: StructOpt + Clone + AugmentClap,
    E: IntoExit,
    RS: FnOnce(E, RunCmd, RP, FactoryFullConfiguration<F>) -> Result<(), String>,
    I: IntoIterator<Item = T>,
    T: Into<OsString> + Clone, 

Parse command line interface arguments and executes the desired command.

Return value

A result that indicates if any error occurred. If no error occurred and a custom subcommand was found, the subcommand is returned. The user needs to handle this subcommand on its own.

Remarks

CC is a custom subcommand. This needs to be an enum! If no custom subcommand is required, NoCustom can be used as type here. RP are custom parameters for the run command. This needs to be a struct! The custom parameters are visible to the user as if they were normal run command parameters. If no custom parameters are required, NoCustom can be used as type here.