Function substrate_cli::parse_and_prepare

pub fn parse_and_prepare<'a, CC, RP, I>(
    version: &'a VersionInfo, 
    impl_name: &'static str, 
    args: I
) -> ParseAndPrepare<'a, CC, RP> where
    CC: StructOpt + Clone + GetLogFilter,
    RP: StructOpt + Clone + AugmentClap,
    I: IntoIterator,
    <I as IntoIterator>::Item: Into<OsString> + Clone, 

Parse command line interface arguments and prepares the command for execution.

Before returning, this function performs various initializations, such as initializing the panic handler and the logger, or increasing the limit for file descriptors.

Remarks

CC is a custom subcommand. This needs to be an enum! If no custom subcommand is required, NoCustom can be used as type here.

RP are custom parameters for the run command. This needs to be a struct! The custom parameters are visible to the user as if they were normal run command parameters. If no custom parameters are required, NoCustom can be used as type here.