Trait substrate_cli::AugmentClap

pub trait AugmentClap: StructOpt {
    fn augment_clap<'a, 'b>(app: App<'a, 'b>) -> App<'a, 'b>;
}

Something that can augment a clap app with further parameters. derive(StructOpt) is implementing this function by default, so a macro impl_augment_clap! is provided to simplify the implementation of this trait.

Required methods

fn augment_clap<'a, 'b>(app: App<'a, 'b>) -> App<'a, 'b>

Augment the given clap App with further parameters.

Loading content...